+++ NEWS.md
@@ -1,5 +1,12 @@
 ## boilerplate (development version)
 
+## [24-08-2024] boilerplate 0.0.0.91
+
+* `boilerplate_manage_measures()`: simple gui to input measures, saves as .rds files 
+* `boilerplate_report_measures()`:  report an appendix of measures with items described.
+* `boilerplate_report_causal_interventions()`: report causal contrasts
+* `boilerplate_report_variables()`:report variables in methods section (exposure/ outcomes)
+
 ## [24-08-2024] boilerplate 0.0.0.9
 
 ### New